refactor: remade the folder architecture, code architecture will follow next
