+++ package.json
@@ -5,21 +5,16 @@
   "dependencies": {
     "@emotion/react": "^11.10.4",
     "@emotion/styled": "^11.10.4",
-    "@mui/material": "^5.10.9",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
-    "@types/axios": "^0.14.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.11.64",
     "@types/react": "^18.0.21",
-    "@types/react-bootstrap": "^0.32.31",
     "@types/react-dom": "^18.0.6",
     "axios": "^1.1.2",
     "dotenv": "^16.0.3",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.5.0",
-    "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
     "react-notifications-component": "^4.0.1",
     "react-router-dom": "^6.4.2",
