+++ src/components/Navbar/index.tsx
@@ -3,8 +3,8 @@ import homeLogo from "../../assets/dashboard/home.png";
 import groupsLogo from "../../assets/dashboard/groups.png";
 import todoLogo from "../../assets/dashboard/todo.png";
 import React from "react";
-import { User } from "../../utils/interfaces";
-import { VIEW } from "../../utils/enum";
+import { User } from "../../domains/utils/interfaces";
+import { VIEW } from "../../domains/utils/enum";
 
 interface NavbarOptions {
     user: User;
