+++ src/domains/pages/home/index.tsx
@@ -1,4 +1,4 @@
-import { Header } from "../../components/Header/Header";
+import { Header } from "../../../components/Header/Header";
 
 export function Home() {
     require("./index.css");
