+++ src/domains/pages/index.tsx
@@ -1,0 +1,5 @@
+export { Dashboard } from "../dashboard";
+export { Home } from "./home";
+export { List } from "./list";
+export { Login, Register } from "../auth/view";
+export { Error404 } from "../error";
