+++ src/domains/utils/enum.ts
