+++ src/domains/utils/interfaces.ts
