--- src/pages/index.tsx
@@ -1,5 +1,0 @@
-export { Dashboard } from "./dashboard";
-export { Home } from "./home";
-export { List } from "./list";
-export { Login, Register } from "./auth";
-export { Error404 } from "./404";
